Handle Garbage Collection of WebGL objects
This is a partial implementation of handling GC of WebGL objects.

What's missing is that in order for it to work the state of all
attachments would have to be tracked. The reason is if you
attach a buffer to vertex array or a texture/renderbuffer to a framebuffer
then even the WebGL object representing that object can get GCed
the actual object is still around (ref counted) inside the driver.

In order to correctly know when an object is truely freed webgl-memory
would therefore have to track all the attachment state so it can
count the references. This issue already existed in webgl-memory
as mentioned in the README as calling `gl.deleteXXX` has the same
issue.

Another thing I don't like about this implementation is it has to
add the id on each WebGL object whereas before the objects were
un-modified. It's probably safe but they do show up in the debugger.

It's slower with this than without as it has to iterate
over all the object `WeakRef`s when you ask for the memory
as there's no way to be notified when an object is collected.

Finally, it's mostly un-testable. As it says on MDN about `WeakRef`
there's no guarantee when GC will happen. The test itself only
passes in Safari as it GCs within about 3 seconds. Chrome and
Firefox take much longer, possibly never.